Fixed banc_codex_annotations
The issue is that the code has changed to do the conversion of the cell_type column into type character after the bind_rows call. It needs to be done before because on the CAVE backend, they have a pandas call that auto figures out what type cell_type should be, and if its value is something that could be a type other than string, it assumes cell_type is that type (e.g. integer). And it depends on more than just where you set the limit and offset, because on the CAVE backend, they grab the table in chunks.